+++ packages/assets-controllers/src/MultichainBalancesController/index.ts
@@ -1,5 +1,6 @@
 export { MultichainBalancesController } from './MultichainBalancesController';
 export type {
+  MultichainAccountBalance,
   MultichainBalancesControllerState,
   MultichainBalancesControllerGetStateAction,
   MultichainBalancesControllerStateChange,